backend/requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.1.0
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
graphene==3.3
graphene-django==3.2.1
graphql-core==3.2.3
graphql-relay==3.2.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
promise==2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dotenv==1.0.1
service-identity==24.1.0
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.0
watchdog==4.0.1
zope.interface==6.4.post2
